Basic Grappling Hook Implemented
The hook is rough but it takes the player from place to place, can hook
onto the AI and allow the player to pull the AI. Melee attacks break the
hook. There is no cooldown currently but one will be added. If you try
to run through the AI it breaks the chain and allows him to melee you.