Massive Update - Hook Fixes/Interface Cooldown
I have fixed most of the poor interactions originally found in the hook.
There is now a cooldown on the hook and an indicator for this cooldown.
The shield has a one time use only indicated by being greyed out
afterward. The yellow attack has been removed from visibility for now
due to issues with it and how it looks aesthetically. The right mouse
button now primarily detaches the hook.